budgetbot/requirements.txt
aiogram==3.18.0
matplotlib==3.8.3
pocketsphinx==0.1.15
pytesseract==0.3.10
Pillow==10.2.0
python-dotenv